+++ api/v2_recommendations_aws_test.go
@@ -1,0 +1,61 @@
+//
+// Author:: Darren Murray (<darren.murray@lacework.net>)
+// Copyright:: Copyright 2022, Lacework Inc.
+// License:: Apache License, Version 2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+package api_test
+
+import (
+	"fmt"
+	"net/http"
+	"testing"
+
+	"github.com/lacework/go-sdk/api"
+	"github.com/lacework/go-sdk/internal/lacework"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestRecommendationsAwsGetReport(t *testing.T) {
+	var (
+		expectedLen = 160
+		fakeServer  = lacework.MockServer()
+	)
+
+	fakeServer.MockToken("TOKEN")
+	fakeServer.UseApiV2()
+	fakeServer.MockAPI("recommendations/aws",
+		func(w http.ResponseWriter, r *http.Request) {
+			assert.Equal(t, "GET", r.Method, "GetReport() should be a GET method")
+			Recommendations := generateRecommendations()
+			fmt.Fprintf(w, Recommendations)
+		},
+	)
+	defer fakeServer.Close()
+
+	c, err := api.NewClient("test",
+		api.WithToken("TOKEN"),
+		api.WithURL(fakeServer.URL()),
+	)
+	assert.Nil(t, err)
+
+	response, err := c.V2.Recommendations.Aws.GetReport("CIS_1_1")
+	assert.Nil(t, err)
+	assert.NotNil(t, response)
+	assert.Equal(t, expectedLen, len(response))
+	for _, rec := range response {
+		assert.NotEmpty(t, rec.ID)
+	}
+}