+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# v1.5.3
+
+## Bug Fixes
+* fix(cli): use request id to fetch scan results (#1110) (Salim Afiune)([f8d633b6](https://github.com/lacework/go-sdk/commit/f8d633b665155611f0ecf512ad3194079eedb679))
+## Other Changes
+* style(cdk): better error message for unsupported env (#1086) (Salim Afiune)([5e21b5f3](https://github.com/lacework/go-sdk/commit/5e21b5f387ee4f3620508e19b88d04ebe5f160d8))
+* chore: Update lacework compliance <csp> help text (#1112) (Ross)([685a4908](https://github.com/lacework/go-sdk/commit/685a49087fe6987ff75939ed20145bd660375593))
+* ci: version bump to v1.5.3-dev (Lacework)([72d4c443](https://github.com/lacework/go-sdk/commit/72d4c4430dcd9ce6a44799313a518d2a12b1f7d6))
+* test(lwupdater): fix windows test (#1113) (Salim Afiune)([0635b838](https://github.com/lacework/go-sdk/commit/0635b8388efb769dc827ca69e6b8c2eb70000938))
+---
+# v1.5.2
+
+## Other Changes
+* chore: re-add compliance enable/disable commands (#1106) (Ross)([f25c72c9](https://github.com/lacework/go-sdk/commit/f25c72c9902bbdcdb22ccf98d28f9e39c57c9dc5))
+* ci: version bump to v1.5.2-dev (Lacework)([9571b913](https://github.com/lacework/go-sdk/commit/9571b91369161598dcb88e8dad22835ea723976e))
+---
 # v1.5.1
 
 ## Bug Fixes